requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.9.2
pyyaml==6.0.2
mlflow==2.16.2
torch==2.4.0
transformers==4.44.2
streamlit==1.39.0
sentence-transformers
fastapi
uvicorn[standard]

